mcp-server/requirements.txt
mcp>=1.0.0
httpx>=0.25.0
pydantic>=2.0.0
uvicorn>=0.24.0
prometheus-client>=0.19.0
python-dotenv>=1.0.0
requests>=2.0.0
